# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Heyarr
  version: "1"
  summary: Headless content lifecycle, replication and consumption for homelabs.
  description: |
    Every error is an RFC 9457 problem document with a stable `type` URI. Clients
    branch on `type` — not on the title, not on the detail, and not on the HTTP
    status alone, which several different conditions share.

    Collections are paginated by keyset with an opaque cursor. `next_cursor` is
    absent on the last page, so "keep going while next_cursor is set" is the
    whole client loop. The cursor's contents are not a contract: it encodes the
    sort key of the last row returned, and a client that parses it breaks the
    first time a sort order changes.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://127.0.0.1:8385
    description: The default loopback bind (ADR-0011).
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/works:
    get:
      tags: [content]
      operationId: listWorks
      summary: Page the catalog.
      description: |
        Ordered by `(sort_title, id)`. The id is in the sort key because
        sort_title is not unique, and a non-unique keyset boundary skips and
        repeats rows exactly as OFFSET does.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: content_type
          in: query
          schema: { type: string }
          description: Restrict to one content type.
        - name: library_id
          in: query
          schema: { type: string }
          description: Works with at least one asset in this library.
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match over `sort_title`.
      responses:
        "200":
          description: A page of works.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Work" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/works/{id}:
    get:
      tags: [content]
      operationId: getWork
      summary: One work.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200":
          description: The work.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Work" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/editions/{id}:
    get:
      tags: [content]
      operationId: getEdition
      summary: One edition.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200":
          description: The edition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Edition" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/assets:
    get:
      tags: [content]
      operationId: listAssets
      summary: Page the files.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: library_id
          in: query
          schema: { type: string }
        - name: content_type
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string, enum: [present, missing] }
          description: |
            `missing` is an asset whose source path has vanished. It is not a
            deletion: the catalog row and any blob both remain (ADR-0018).
      responses:
        "200":
          description: A page of assets.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/assets/{id}:
    get:
      tags: [content]
      operationId: getAsset
      summary: One asset.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200":
          description: The asset.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [content]
      operationId: deleteAsset
      summary: Delete an asset. Requires `write`.
      description: |
        Logical, per ADR-0018. The asset stops existing; its blob does not. A
        later `gc_blobs` sweep reclaims blobs that no asset references, once a
        grace window has passed — which is what makes a mistaken delete
        reversible. Nothing is unlinked inside this request.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204":
          description: Deleted. No bytes were removed.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/libraries:
    get:
      tags: [content]
      operationId: listLibraries
      summary: Page the libraries, each with its roots.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: content_type
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match over the name.
      responses:
        "200":
          description: A page of libraries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Library" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [content]
      operationId: createLibrary
      summary: Create a library. Requires `write`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content_type]
              additionalProperties: false
              properties:
                name: { type: string }
                content_type: { type: string }
                enabled: { type: boolean, default: true }
      responses:
        "201":
          description: Created.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Library" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/libraries/{id}:
    get:
      tags: [content]
      operationId: getLibrary
      summary: One library, with its roots.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200":
          description: The library.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Library" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/libraries/{id}/roots:
    post:
      tags: [content]
      operationId: createLibraryRoot
      summary: Add a root to a library. Requires `write`.
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              additionalProperties: false
              properties:
                path: { type: string }
                ingest_mode:
                  type: string
                  enum: [reflink, hardlink, copy, link]
                  default: reflink
                  description: |
                    How ingest materialises bytes into the CAS (ADR-0014).
                    `link` catalogues files where they are and never copies them
                    (ADR-0020).
                enabled: { type: boolean, default: true }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryRoot" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/libraries/{id}/scan:
    post:
      tags: [content]
      operationId: scanLibrary
      summary: Enqueue a scan. Requires `write`.
      description: |
        One `scan_library` job per enabled root, because the scanner's job is
        root-scoped. A library with two roots therefore yields two jobs, and a
        client waiting on the scan must wait on all of them — waiting on the
        first only is how a CLI exits 0 with half the library unscanned.

        Enqueueing is idempotent while a scan of the same root is still live: a
        second request returns the running job rather than queueing a second
        walk of the same tree (ADR-0008).
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "202":
          description: Accepted; the jobs are queued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  library_id: { type: string }
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            The library has no enabled roots, so there is nothing to scan.
            Returning 202 here would let a client believe a scan is running.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/blobs/{hash}:
    get:
      tags: [bytes]
      operationId: getBlob
      summary: Blob metadata.
      description: |
        Metadata only. The bytes are served by `/api/v1/blobs/{hash}/content`,
        which is deliberately a separate contract with four consumers
        (ADR-0013).
      parameters:
        - name: hash
          in: path
          required: true
          schema: { type: string, pattern: "^blake3:[0-9a-f]{64}$" }
      responses:
        "200":
          description: The blob record.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Blob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/renderers:
    get:
      tags: [playback]
      operationId: listRenderers
      summary: List the renderers on this node's network (§68).
      description: >-
        Televisions, speakers and projectors that Heyarr can play to, found by
        SSDP on the local segment.


        An empty list means nothing ANSWERED, not that nothing is there. A
        screen in standby closes every listener and on some models leaves the
        network entirely, so a device that is switched off is indistinguishable
        from one that does not exist. Discovery is also multicast, so it does
        not cross a routed link: a controller on the far side of a VPN finds
        nothing at all.
      parameters:
        - name: refresh
          in: query
          required: false
          description: >-
            `true` searches again instead of reusing the cached sweep. A search
            takes several seconds, so this is for "I just switched it on".
          schema: { type: string }
      responses:
        '200':
          description: What answered.
        '401': { $ref: "#/components/responses/Unauthorized" }
  /api/v1/renderers/{udn}/status:
    parameters:
      - name: udn
        in: path
        required: true
        description: The renderer's UPnP UDN, from `GET /api/v1/renderers`.
        schema: { type: string }
    get:
      tags: [playback]
      operationId: getRendererStatus
      summary: Report what a renderer is doing.
      description: >-
        Transport state and, when the device reports one, position. A renderer
        that answers no position is not broken — `NOT_IMPLEMENTED` is a legal
        answer and some devices give it forever.
      responses:
        '200':
          description: The renderer's state.
        '404':
          description: No renderer with that id answered.
        '502':
          description: The renderer did not answer.
...
  /api/v1/renderers/{udn}/pause:
    parameters:
      - name: udn
        in: path
        required: true
        description: The renderer's UPnP UDN, from `GET /api/v1/renderers`.
        schema: { type: string }
    post:
      tags: [playback]
      operationId: pauseRenderer
      summary: Hold position.
      responses:
        '200': { description: The renderer's state after pausing. }
        '502': { description: The renderer refused. }
  /api/v1/renderers/{udn}/resume:
    parameters:
      - name: udn
        in: path
        required: true
        description: The renderer's UPnP UDN, from `GET /api/v1/renderers`.
        schema: { type: string }
    post:
      tags: [playback]
      operationId: resumeRenderer
      summary: Continue from where it was paused.
      description: >-
        This is UPnP's Play. There is no separate resume verb: Play from
        PAUSED_PLAYBACK continues, and Play from STOPPED restarts.
      responses:
        '200': { description: The renderer's state after resuming. }
        '502': { description: The renderer refused. }
  /api/v1/renderers/{udn}/stop:
    parameters:
      - name: udn
        in: path
        required: true
        description: The renderer's UPnP UDN, from `GET /api/v1/renderers`.
        schema: { type: string }
    post:
      tags: [playback]
      operationId: stopRenderer
      summary: End playback and release the content.
      responses:
        '200': { description: The renderer's state after stopping. }
        '502': { description: The renderer refused. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An opaque scoped token (ADR-0011), presented as `Authorization: Bearer
        heyarr_…`. The header is the only place a credential is accepted: a
        `?token=` parameter would write it into every proxy log, browser history
        and Referer header between the client and here.

        Scopes are ordered — `admin` implies `write` implies `read`. Every
        route under /api/v1 requires at least `read`.

    peerCertificate:
      type: mutualTLS
      description: |
        A self-signed X.509 certificate carrying this peer's Ed25519 public
        key, pinned against the controller-issued membership record
        (ADR-0012). There is no CA and no PKI: the certificate proves
        possession of a key, and membership decides whether that key may
        speak. Both ends pin, so the dialler authenticates the listener by
        exactly the same rule.

        It is checked twice — during the handshake, and again on every request
        the connection carries. Revocation is deleting the membership record,
        and there is no revocation list to fall back on, so the freshness of
        that second lookup is the mechanism.

        This credential is accepted ONLY on /peer/v1, and a bearer token is
        never accepted there. A connection presenting no client certificate is
        refused during the handshake, not with a status code.
...